fpc-backend/requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
exceptiongroup==1.3.0
fastapi==0.120.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
ultralytics==8.0.196
